server/requirements.txt
python-dotenv
psycopg2
fastapi[standard]
